fix: move utils to the correct place
