+++ src/faker.ts
@@ -28,9 +28,9 @@ import { ScienceModule } from './modules/science';
 import { SystemModule } from './modules/system';
 import { VehicleModule } from './modules/vehicle';
 import { WordModule } from './modules/word';
+import type { Randomizer } from './randomizer';
 import { SimpleFaker } from './simple-faker';
 import { mergeLocales } from './utils/merge-locales';
-import type { Randomizer } from './utils/randomizer';
 
 /**
  * This is Faker's main class containing all modules that can be used to generate data.
